feat: Player 기능 구현
- 플레이어 기준으로 승패를 결정하게 변경함
