+++ README.md
@@ -11,12 +11,14 @@
 - [x]  복수의 카드 숫자를 합하는 기능
 - [x]  승무패 계산 기능
 - [x]  참여자에게 카드를 나눠주는 기능
-- [ ]  카드를 중복으로 뽑지 않게 하는 기능
+- [x]  카드를 중복으로 뽑지 않게 하는 기능
 - [x]  Ace 카드를 1또는 11로 계산하게 하는 기능
 - [x]  숫자가 특정 값 이하인지 판단하는 기능
 - [x]  딜러인 경우 16
 - [x]  플레이어인 경우 21
 - [x]  카드를 더 받을 수 있는지 확인하는 기능
+- [ ]  게임 결과를 저장하는 기능
+- [x]  딜러의 경우, 가장 첫 카드 한장만 보여주는 기능
 
 # 게임 플로우
