+++ src/test/java/domain/DealerTest.java
@@ -1,8 +1,5 @@
 package domain;
 
-import static domain.GameResult.DRAW;
-import static domain.GameResult.LOSE;
-import static domain.GameResult.WIN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
@@ -19,36 +16,11 @@ class DealerTest {
 
     @ParameterizedTest
     @MethodSource
-    @DisplayName("승무패를 계산하는지 확인합니다.")
-    void decideGameResultTest(List<Card> cards, GameResult expected) {
-        //given
-        Player player = new Player("ㅎㅎ");
-        cards.forEach(player::addCard);
-        Dealer dealer = new Dealer(new Deck(new ArrayList<>()));
-        dealer.addCard(Card.HEART_JACK);
-        dealer.addCard(Card.SPADE_QUEEN); // 20
-
-        //when & then
-        Assertions.assertEquals(expected, dealer.decideGameResult(player));
-    }
-
-    public static Stream<Arguments> decideGameResultTest() {
-        return Stream.of(
-                Arguments.of(List.of(Card.HEART_ACE, Card.CLOVER_JACK, Card.CLOVER_THREE), WIN),
-                Arguments.of(List.of(Card.HEART_ACE, Card.CLOVER_JACK, Card.CLOVER_QUEEN), LOSE),
-                Arguments.of(List.of(Card.HEART_ACE, Card.CLOVER_JACK, Card.CLOVER_ACE), WIN),
-                Arguments.of(List.of(Card.HEART_SEVEN, Card.CLOVER_JACK, Card.CLOVER_THREE), DRAW),
-                Arguments.of(List.of(Card.HEART_JACK, Card.CLOVER_JACK, Card.CLOVER_ACE, Card.HEART_ACE), WIN)
-        );
-    }
-
-    @ParameterizedTest
-    @MethodSource
     @DisplayName("딜러가 카드를 더 받을 수 있는지 확인합니다.")
     void canGetMoreCardTest(List<Card> cards, boolean expected) {
         //given
         Dealer dealer = new Dealer(new Deck(new ArrayList<>()));
-        cards.forEach(dealer::addCard);
+        dealer.addCard(cards);
 
         //when & then
         assertThat(dealer.canGetMoreCard()).isEqualTo(expected);
@@ -72,7 +44,7 @@ class DealerTest {
     void showFirstCardTest(List<Card> cards) {
         //given
         Dealer dealer = new Dealer(new Deck(new ArrayList<>()));
-        cards.forEach(dealer::addCard);
+        dealer.addCard(cards);
         //when & then
         org.assertj.core.api.Assertions.assertThat(cards).contains(dealer.showAnyOneCard());
     }
